feat! Close #1 android setup
