+++ android/build.gradle
@@ -1,4 +1,4 @@
-allprojects {
+pallprojects {
     repositories {
         google()
         mavenCentral()
@@ -6,11 +6,10 @@ allprojects {
 }
 
 rootProject.buildDir = '../build'
-subprojects {
-    project.buildDir = "${rootProject.buildDir}/${project.name}"
-}
+
 subprojects {
     project.evaluationDependsOn(':app')
+    project.buildDir = "${rootProject.buildDir}/${project.name}"
 }
 
 tasks.register("clean", Delete) {
