+++ lib/counter/view/counter_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:instagram/counter/counter.dart';
+import 'package:instagram/counter/view/widgets/widgets.dart';
 import 'package:instagram/l10n/l10n.dart';
 
 class CounterPage extends StatelessWidget {
@@ -40,16 +41,5 @@ class CounterView extends StatelessWidget {
         ],
       ),
     );
-  }
-}
-
-class CounterText extends StatelessWidget {
-  const CounterText({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final count = context.select((CounterCubit cubit) => cubit.state);
-    return Text('$count', style: theme.textTheme.displayLarge);
   }
 }
