+++ lib/counter/view/widgets/counter_text.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:instagram/counter/cubit/counter_cubit.dart';
+
+class CounterText extends StatelessWidget {
+  const CounterText({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final count = context.select((CounterCubit cubit) => cubit.state);
+    return Text('$count', style: theme.textTheme.displayLarge);
+  }
+}
