--- lib/l10n/arb/app_localizations_es.dart
@@ -1,9 +1,0 @@
-import 'app_localizations.dart';
-
-/// The translations for Spanish Castilian (`es`).
-class AppLocalizationsEs extends AppLocalizations {
-  AppLocalizationsEs([String locale = 'es']) : super(locale);
-
-  @override
-  String get counterAppBarTitle => 'Contador';
-}
