+++ pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   intl: ^0.19.0
 
   flutter_gen: ^5.6.0
+  provider: ^6.1.2
 dev_dependencies:
   bloc_test: ^9.1.7
   flutter_test:
